refactor(friends): plier la page amis dans la liste
`/friends` affichait un fil des activités d'amis — exactement les cartes que la
liste montre déjà sous sa portée « Amis » depuis la fusion — plus trois actions
qui, elles, n'existaient nulle part ailleurs : montrer son QR, en scanner un, et
gérer la liste. Le fil part, les trois actions restent.

Elles suivent désormais la portée qui les réclame : sélectionner « Amis » sur
la liste les fait apparaître au-dessus des cartes. `FriendsQuickActions` est
autonome, modales comprises — l'hôte le rend et ne lui doit rien.

Du coup les onglets de portée s'affichent toujours, y compris sans personne à
suivre. Ils étaient conditionnés à l'existence d'un ami, ce qui se tenait quand
il n'y avait rien à choisir ; ça ne tient plus maintenant que l'onglet « Amis »
est la seule porte pour en ajouter un. La masquer jusqu'à ce qu'on ait déjà un
ami, c'est la fermer à ceux qui en cherchent.

Sous cette portée, une liste vide nomme ce qui manque au lieu de répéter les
boutons qui sont juste au-dessus.

`/friends` redirige vers `/?who=friends`, pour les liens partagés et pour
l'entrée du menu. `useFriendsFeed` disparaît avec sa vue : il n'avait pas
d'autre appelant.
