--- src/composables/useFriendsFeed.ts
@@ -1,115 +1,0 @@
-import { ref, computed } from 'vue'
-import { ActivityFeedService, type FeedActivity } from '@/services/ActivityFeedService'
-
-// Re-export FeedActivity type for convenience
-export type { FeedActivity }
-
-/**
- * Composable for managing friends-only activity feed state
- * Filters activities to show only friend activities
- * Handles pagination, loading states, and UI interactions
- * Delegates data loading and transformation to ActivityFeedService
- */
-export function useFriendsFeed() {
-  // State
-  const activities = ref<FeedActivity[]>([])
-  const loading = ref(false)
-  const hasMore = ref(true)
-  const page = ref(0)
-  const pageSize = 10
-
-  // Cached data and concurrency control. A ref, not a plain variable: `count` is
-  // computed from it, and a computed over a non-reactive local never updates —
-  // the counter it feeds was permanently stuck at zero.
-  const allActivities = ref<FeedActivity[]>([])
-  let loadMorePromise: Promise<void> | null = null
-
-  // Service dependency
-  const feedService = ActivityFeedService.getInstance()
-
-  /**
-   * Load all activities from service and filter friends only
-   */
-  const loadAllActivities = async (): Promise<FeedActivity[]> => {
-    const all = await feedService.loadAllActivities()
-    // Filter to keep only friend activities
-    allActivities.value = all.filter(activity => activity.source === 'friend')
-    return allActivities.value
-  }
-
-  /**
-   * Load next page of activities
-   */
-  const loadMore = async () => {
-    // If already loading, wait for existing operation to complete
-    if (loadMorePromise) {
-      return loadMorePromise
-    }
-
-    // Early return if no more data
-    if (!hasMore.value) return
-
-    // Create and store the loading promise
-    loadMorePromise = (async () => {
-      loading.value = true
-
-      try {
-        // Load all activities if not loaded yet
-        if (allActivities.value.length === 0) {
-          await loadAllActivities()
-        }
-
-        // Calculate pagination
-        const start = page.value * pageSize
-        const end = start + pageSize
-        const newActivities = allActivities.value.slice(start, end)
-
-        if (newActivities.length < pageSize) {
-          hasMore.value = false
-        }
-
-        activities.value.push(...newActivities)
-        page.value += 1
-      } catch (error) {
-        console.error('[useFriendsFeed] Error loading activities:', error)
-      } finally {
-        loading.value = false
-        loadMorePromise = null
-      }
-    })()
-
-    return loadMorePromise
-  }
-
-  /**
-   * Reload feed from scratch (e.g., after refresh)
-   */
-  const reload = async () => {
-    // Wait for any ongoing load to complete
-    if (loadMorePromise) {
-      await loadMorePromise
-    }
-
-    activities.value = []
-    allActivities.value = []
-    page.value = 0
-    hasMore.value = true
-    loadMorePromise = null
-
-    await loadMore()
-  }
-
-  /**
-   * Get count of friend activities
-   */
-  const count = computed(() => allActivities.value.length)
-
-  return {
-    activities,
-    loading,
-    hasMore,
-    loadMore,
-    reload,
-    count
-  }
-}
