+++ src/locales/fr.json
@@ -283,13 +283,9 @@
     "minutes": "min"
   },
   "friends": {
-    "title": "Activités des amis",
-    "activities": "Activités",
-    "sync": "Synchroniser",
     "addFriend": "Ajouter un ami",
     "noActivities": "Aucune activité d'amis",
     "noActivitiesDescription": "Ajoutez des amis pour voir leurs activités ici",
-    "manageFriends": "Gérer mes amis",
     "actions": "Actions amis",
     "quickQr": "Mon QR",
     "quickAdd": "Ajouter",
